+++ actionview/lib/action_view/gem_version.rb
@@ -10,7 +10,7 @@ module ActionView
     MAJOR = 8
     MINOR = 1
     TINY  = 0
-    PRE   = "alpha"
+    PRE   = "beta1"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end