+++ activejob/lib/active_job.rb
@@ -43,6 +43,7 @@ module ActiveJob
   autoload :EnqueueAfterTransactionCommit
 
   eager_autoload do
+    autoload :Continuable
     autoload :Continuation
     autoload :Serializers
     autoload :ConfiguredJob