+++ activerecord/test/cases/adapter_test.rb
@@ -882,13 +882,13 @@ module ActiveRecord
       end
 
       test "#execute is retryable" do
-        initial_connection_id = connection_id_from_server
+        initial_connection_id = connection_id_from_server(@connection)
 
         kill_connection_from_server(initial_connection_id)
 
         @connection.execute("SELECT 1", allow_retry: true)
 
-        assert_not_equal initial_connection_id, connection_id_from_server
+        assert_not_equal initial_connection_id, connection_id_from_server(@connection)
       end
 
       test "disconnect and recover on #configure_connection failure" do
@@ -930,79 +930,10 @@ module ActiveRecord
         end
 
         assert_equal [[1]], connection.exec_query("SELECT 1").rows
-        assert_empty failures
+        assert_empty slow
       ensure
         connection&.disconnect!
       end
-
-      private
-        def raw_transaction_open?(connection)
-          case connection.adapter_name
-          when "PostgreSQL"
-            connection.instance_variable_get(:@raw_connection).transaction_status == ::PG::PQTRANS_INTRANS
-          when "Mysql2", "Trilogy"
-            begin
-              connection.instance_variable_get(:@raw_connection).query("SAVEPOINT transaction_test")
-              connection.instance_variable_get(:@raw_connection).query("RELEASE SAVEPOINT transaction_test")
-
-              true
-            rescue
-              false
-            end
-          when "SQLite"
-            begin
-              connection.instance_variable_get(:@raw_connection).transaction { nil }
-              false
-            rescue
-              true
-            end
-          else
-            skip("raw_transaction_open? unsupported")
-          end
-        end
-
-        def remote_disconnect(connection)
-          case connection.adapter_name
-          when "PostgreSQL"
-            # Connection was left in a bad state, need to reconnect to simulate fresh disconnect
-            connection.verify! if connection.instance_variable_get(:@raw_connection).status == ::PG::CONNECTION_BAD
-            unless connection.instance_variable_get(:@raw_connection).transaction_status == ::PG::PQTRANS_INTRANS
-              connection.instance_variable_get(:@raw_connection).async_exec("begin")
-            end
-            connection.instance_variable_get(:@raw_connection).async_exec("set idle_in_transaction_session_timeout = '10ms'")
-            sleep 0.05
-          when "Mysql2", "Trilogy"
-            connection.send(:internal_execute, "set @@wait_timeout=1", materialize_transactions: false)
-            sleep 1.2
-          else
-            skip("remote_disconnect unsupported")
-          end
-        end
-
-        def connection_id_from_server
-          case @connection.adapter_name
-          when "Mysql2", "Trilogy"
-            @connection.execute("SELECT CONNECTION_ID()").to_a[0][0]
-          when "PostgreSQL"
-            @connection.execute("SELECT pg_backend_pid()").to_a[0]["pg_backend_pid"]
-          else
-            skip("connection_id_from_server unsupported")
-          end
-        end
-
-        def kill_connection_from_server(connection_id)
-          conn = @connection.pool.checkout
-          case conn.adapter_name
-          when "Mysql2", "Trilogy"
-            conn.execute("KILL #{connection_id}")
-          when "PostgreSQL"
-            conn.execute("SELECT pg_terminate_backend(#{connection_id})")
-          else
-            skip("kill_connection_from_server unsupported")
-          end
-
-          conn.close
-        end
     end
   end