+++ activesupport/lib/active_support/railtie.rb
@@ -41,7 +41,7 @@ module ActiveSupport
     initializer "active_support.set_event_reporter_context_store" do |app|
       config.after_initialize do
         if klass = app.config.active_support.event_reporter_context_store
-          ActiveSupport.event_reporter.context_store = klass
+          ActiveSupport::EventReporter.context_store = klass
         end
       end
     end
@@ -76,6 +76,12 @@ module ActiveSupport
           require "active_support/execution_context/test_helper"
           include ActiveSupport::ExecutionContext::TestHelper
         end
+      end
+    end
+
+    initializer "active_support.set_filter_parameters" do |app|
+      config.after_initialize do
+        ActiveSupport.filter_parameters += Rails.application.config.filter_parameters
       end
     end