+++ activemodel/CHANGELOG.md
@@ -8,6 +8,9 @@
 
     *Jevin Sew*
 
+
+## Rails 8.1.0.beta1 (September 04, 2025) ##
+
 *   Add `except_on:` option for validation callbacks.
 
     *Ben Sheldon*