+++ activestorage/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Rails 8.1.0.beta1 (September 04, 2025) ##
+
+*   Remove deprecated `:azure` storage service.
+
+    *Rafael Mendonça França*
+
 *   Remove unnecessary calls to the GCP metadata server.
 
     Calling Google::Auth.get_application_default triggers an explicit call to